Move around imports, stub out test.
